requirements.txt
mysql-connector-python
pyspark
kafka-python